* 종스크롤 형태의 loop Background 구현 * 되는것 확인만 한 상태, 추가 코드 정리 필요

+++ ShootingStrike/HammerEffect.cpp
@@ -53,15 +53,15 @@ int HammerEffect::Update()
 void HammerEffect::Render(HDC _hdc)
 {
 	TransparentBlt(_hdc, // ** 최종 출력 위치
-		TransInfo.Position.x - (TransInfo.Scale.x / 2) + Offset.x,
-		TransInfo.Position.y - (TransInfo.Scale.y / 2) + Offset.y,
-		TransInfo.Scale.x,
-		TransInfo.Scale.y,
+		(int)TransInfo.Position.x - (int)(TransInfo.Scale.x / 2) + (int)Offset.x,
+		(int)TransInfo.Position.y - (int)(TransInfo.Scale.y / 2) + (int)Offset.y,
+		(int)TransInfo.Scale.x,
+		(int)TransInfo.Scale.y,
 		BitmapManager::GetInstance()->GetMemDC(eImageKey::EFFECT),
-		TransInfo.Scale.x * Frame,
-		TransInfo.Scale.y * 0,
-		TransInfo.Scale.x,
-		TransInfo.Scale.y,
+		(int)TransInfo.Scale.x * Frame,
+		(int)TransInfo.Scale.y * 0,
+		(int)TransInfo.Scale.x,
+		(int)TransInfo.Scale.y,
 		RGB(255, 0, 255));
 }
